+++ Sources/Sentry/SentryEvent.m
@@ -118,7 +118,7 @@ SentryEvent ()
 
 - (void)addSimpleProperties:(NSMutableDictionary *)serializedData
 {
-    [serializedData setValue:self.sdk forKey:@"sdk"];
+    [serializedData setValue:[self.sdk sentry_sanitize] forKey:@"sdk"];
     [serializedData setValue:self.releaseName forKey:@"release"];
     [serializedData setValue:self.dist forKey:@"dist"];
     [serializedData setValue:self.environment forKey:@"environment"];
@@ -138,7 +138,7 @@ SentryEvent ()
 
     [serializedData setValue:[self serializeBreadcrumbs] forKey:@"breadcrumbs"];
 
-    [serializedData setValue:self.context forKey:@"contexts"];
+    [serializedData setValue:[self.context sentry_sanitize] forKey:@"contexts"];
 
     if (nil != self.message) {
         [serializedData setValue:[self.message serialize] forKey:@"message"];