+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Enhance the UIViewController breadcrumbs with more data (#1945)
 
+### Fixes
+
+- Properly sanitize the event context and SDK information (#1943)
+
 ## 7.20.0
 
 ### Features